# openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/centers/{centerId}/services:
    get:
      operationId: services_list
      parameters:
        - name: X-Center-Id
          in: header
          required: true
          schema:
            type: string
        - name: centerId
          required: true
          in: path
          schema:
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceListResponseDto"
        default:
          description: Error RFC 9457
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetailsDto"
      security:
        - bearer: []
      summary: Servicios del centro. La clientela solo ve los visibles; el equipo ve todos los no archivados.
      tags:
        - services
    post:
      operationId: services_create
      parameters:
        - name: X-Center-Id
          in: header
          required: true
          schema:
            type: string
        - name: centerId
          required: true
          in: path
          schema:
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequestDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponseDto"
        default:
          description: Error RFC 9457
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetailsDto"
      security:
        - bearer: []
      summary: Crea un servicio individual. Si no se indica quién lo atiende, lo atiende quien lo crea.
      tags:
        - services
  /v1/centers/{centerId}/services/{serviceId}:
    patch:
      operationId: services_update
      parameters:
        - name: X-Center-Id
          in: header
          required: true
          schema:
            type: string
        - name: centerId
          required: true
          in: path
          schema:
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            type: string
        - name: serviceId
          required: true
          in: path
          schema:
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceRequestDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponseDto"
        default:
          description: Error RFC 9457
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetailsDto"
      security:
        - bearer: []
      summary: Modifica un servicio. Solo cambia lo que se envía; null borra descripción, precio o color.
      tags:
        - services
    delete:
      operationId: services_archive
      parameters:
        - name: X-Center-Id
          in: header
          required: true
          schema:
            type: string
        - name: centerId
          required: true
          in: path
          schema:
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            type: string
        - name: serviceId
          required: true
          in: path
          schema:
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            type: string
      responses:
        "204":
          description: ""
        default:
          description: Error RFC 9457
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetailsDto"
      security:
        - bearer: []
      summary: "Archiva el servicio: deja de ofrecerse, pero las reservas existentes se conservan."
      tags:
        - services
...
  /v1/centers/{centerId}/bookings:
    post:
      operationId: bookings_create
      parameters:
        - name: X-Center-Id
          in: header
          required: true
          schema:
            type: string
        - name: centerId
          required: true
          in: path
          schema:
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            type: string
        - name: Idempotency-Key
          in: header
          description: UUID por intención del usuario; los reintentos reutilizan la misma
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequestDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponseDto"
        default:
          description: Error RFC 9457
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetailsDto"
      security:
        - bearer: []
      summary: "Reserva un hueco individual. La hora debe coincidir con un hueco de /availability: 409 SLOT_UNAVAILABLE si ya no está, OUTSIDE_BOOKING_WINDOW si es demasiado pronto o lejos, ALREADY_BOOKED si ya tienes otra cita a esa hora. Repetir la misma Idempotency-Key devuelve la misma reserva."
      tags:
        - bookings
  /v1/centers/{centerId}/bookings/mine:
    get:
      operationId: bookings_list_mine
      parameters:
        - name: X-Center-Id
          in: header
          required: true
          schema:
            type: string
        - name: centerId
          required: true
          in: path
          schema:
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            type: string
        - name: scope
          required: false
          in: query
          schema:
            default: upcoming
            type: string
            enum:
              - upcoming
              - past
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 50
            default: 20
            type: integer
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyBookingsResponseDto"
        default:
          description: Error RFC 9457
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetailsDto"
      security:
        - bearer: []
      summary: "Mis reservas. «upcoming»: no canceladas que aún no han terminado, de la más próxima a la más lejana. «past»: el resto, de la más reciente a la más antigua."
      tags:
        - bookings
  /v1/centers/{centerId}/bookings/{bookingId}/cancel:
    post:
      operationId: bookings_cancel
      parameters:
        - name: X-Center-Id
          in: header
          required: true
          schema:
            type: string
        - name: centerId
          required: true
          in: path
          schema:
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            type: string
        - name: bookingId
          required: true
          in: path
          schema:
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelBookingResponseDto"
        default:
          description: Error RFC 9457
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetailsDto"
      security:
        - bearer: []
      summary: Cancela una reserva propia. Se puede cancelar fuera de plazo (withinPolicy=false). Cancelar una ya cancelada devuelve el mismo resultado; una que ya empezó, 409 BOOKING_NOT_CANCELLABLE.
      tags:
        - bookings
...
info:
  title: Yoclick API
  description: API multi-centro de reservas en marca blanca. Contrato canónico para la app.
  version: 0.1.0
  contact: {}
...
servers:
  - url: https://api.yoclick.app
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    ProblemDetailsDto:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        code:
          type: string
        traceId:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              code:
                type: string
            required:
              - path
              - code
      required:
        - type
        - title
        - status
        - code
        - traceId
...
    ServiceListResponseDto:
      type: object
      properties:
        services:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
                pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
              name:
                type: string
              description:
                type: string
                nullable: true
              kind:
                type: string
                enum:
                  - individual
              durationMinutes:
                type: integer
                minimum: -9007199254740991
                maximum: 9007199254740991
              priceCents:
                type: integer
                minimum: -9007199254740991
                maximum: 9007199254740991
                nullable: true
              currency:
                type: string
                enum:
                  - EUR
              color:
                type: string
                nullable: true
              bookingWindowDays:
                type: integer
                minimum: -9007199254740991
                maximum: 9007199254740991
              minNoticeMinutes:
                type: integer
                minimum: -9007199254740991
                maximum: 9007199254740991
              isVisible:
                type: boolean
              staff:
                type: array
                items:
                  type: object
                  properties:
                    membershipId:
                      type: string
                      format: uuid
                      pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                    fullName:
                      type: string
                  required:
                    - membershipId
                    - fullName
                  additionalProperties: false
            required:
              - id
              - name
              - description
              - kind
              - durationMinutes
              - priceCents
              - currency
              - color
              - bookingWindowDays
              - minNoticeMinutes
              - isVisible
              - staff
            additionalProperties: false
      required:
        - services
      additionalProperties: false
    CreateServiceRequestDto:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
        description:
          type: string
          minLength: 1
          maxLength: 500
        durationMinutes:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        priceCents:
          type: integer
          minimum: 0
          maximum: 10000000
        color:
          type: string
          pattern: ^#[0-9A-Fa-f]{6}$
        bookingWindowDays:
          type: integer
          minimum: 1
          maximum: 365
        minNoticeMinutes:
          type: integer
          minimum: 0
          maximum: 86400
        isVisible:
          type: boolean
        staffMembershipIds:
          minItems: 1
          maxItems: 50
          type: array
          items:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
      required:
        - name
        - durationMinutes
      additionalProperties: false
    ServiceResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        name:
          type: string
        description:
          type: string
          nullable: true
        kind:
          type: string
          enum:
            - individual
        durationMinutes:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        priceCents:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
          nullable: true
        currency:
          type: string
          enum:
            - EUR
        color:
          type: string
          nullable: true
        bookingWindowDays:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        minNoticeMinutes:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        isVisible:
          type: boolean
        staff:
          type: array
          items:
            type: object
            properties:
              membershipId:
                type: string
                format: uuid
                pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
              fullName:
                type: string
            required:
              - membershipId
              - fullName
            additionalProperties: false
      required:
        - id
        - name
        - description
        - kind
        - durationMinutes
        - priceCents
        - currency
        - color
        - bookingWindowDays
        - minNoticeMinutes
        - isVisible
        - staff
      additionalProperties: false
    UpdateServiceRequestDto:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
        description:
          type: string
          minLength: 1
          maxLength: 500
          nullable: true
        durationMinutes:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        priceCents:
          type: integer
          minimum: 0
          maximum: 10000000
          nullable: true
        color:
          type: string
          pattern: ^#[0-9A-Fa-f]{6}$
          nullable: true
        bookingWindowDays:
          type: integer
          minimum: 1
          maximum: 365
        minNoticeMinutes:
          type: integer
          minimum: 0
          maximum: 86400
        isVisible:
          type: boolean
        staffMembershipIds:
          minItems: 1
          maxItems: 50
          type: array
          items:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
      additionalProperties: false
...
    CreateBookingRequestDto:
      type: object
      properties:
        serviceId:
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        startsAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z))$
        staffMembershipId:
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
      required:
        - serviceId
        - startsAt
      additionalProperties: false
    BookingResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        status:
          type: string
          enum:
            - confirmed
            - cancelled
            - attended
            - no_show
        startsAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z))$
        endsAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z))$
        service:
          type: object
          properties:
            id:
              type: string
              format: uuid
              pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            name:
              type: string
            durationMinutes:
              type: integer
              minimum: -9007199254740991
              maximum: 9007199254740991
            color:
              type: string
              nullable: true
          required:
            - id
            - name
            - durationMinutes
            - color
          additionalProperties: false
        staff:
          type: object
          properties:
            membershipId:
              type: string
              format: uuid
              pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            fullName:
              type: string
          required:
            - membershipId
            - fullName
          additionalProperties: false
        cancelledAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z))$
          nullable: true
        cancelWithinPolicy:
          type: boolean
          nullable: true
        createdAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z))$
      required:
        - id
        - status
        - startsAt
        - endsAt
        - service
        - staff
        - cancelledAt
        - cancelWithinPolicy
        - createdAt
      additionalProperties: false
...